backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic==2.9.0
pydantic-settings==2.5.0
paho-mqtt==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.0
openai==1.50.0
pulp==2.9.0
websockets==13.0
alembic==1.13.0
